% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSS.R
\name{BSS}
\alias{BSS}
\title{BSS Function}
\usage{
BSS(x, y, priorValue, runs, burn)
}
\arguments{
\item{x}{The set of predictor variables}

\item{y}{The response variable}

\item{priorValue}{test}

\item{runs}{test}

\item{burn}{test}
}
\value{
Returns a list
}
\description{
BSS Function
}
