% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat}
\alias{dat}
\title{Example dataset for \code{ssvs} function
@format A data frame with 74 records and 76 variables}
\format{
An object of class \code{data.frame} with 74 rows and 76 columns.
}
\usage{
dat
}
\description{
Example dataset for \code{ssvs} function
@format A data frame with 74 records and 76 variables
}
\keyword{datasets}
