% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputed_affairs.R
\docType{data}
\name{imputed_affairs}
\alias{imputed_affairs}
\title{Imputed affairs Dataset}
\format{
A data frame with 3005 rows and 12 variables
}
\source{
Original dataset from \code{datasets::Affairs}, with missing values introduced and imputed.
}
\usage{
imputed_affairs
}
\description{
This dataset is a version of the \code{Affairs} dataset where random missing values
were introduced, and multiple imputation was performed using the \code{mice} package.
}
\details{
Random missingness was introduced into 10\% of the values in the original \code{Affairs} dataset.
Multiple imputation was then performed using the \code{mice} package with the following parameters:
\itemize{
\item 5 multiple imputations (\code{m = 5}).
\item 50 iterations per imputation (\code{maxit = 50}).
\item Seed set to 123 for reproducibility.
}
The dataset included here is the first completed dataset resulting from the multiple imputation process.
}
\examples{
\donttest{
data(imputed_affairs)
head(imputed_affairs)
}
}
\keyword{datasets}
