% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch}
\alias{launch}
\title{Run an interactive analysis tool (Shiny app) that lets you perform SSVS in a browser}
\usage{
launch()
}
\description{
Run an interactive analysis tool (Shiny app) that lets you perform SSVS in a browser
}
