% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MI.R
\name{plot.est}
\alias{plot.est}
\title{Plot SSVS-MI Estimates with Confidence Intervals}
\usage{
\method{plot}{est}(data, title = NULL)
}
\arguments{
\item{data}{A data frame containing the summary statistics for SSVS results. Must include columns
\code{Variables}, \code{avg.beta}, \code{min}, and \code{max} with option \code{Condition}.}

\item{ty}{A character vector specifying the order of conditions in the plot.}

\item{pal}{A character vector of colors corresponding to the conditions.}
}
\value{
A \code{ggplot2} object representing the SSVS estimates plot.
}
\description{
This function creates a plot of SSVS-MI estimates with confidence intervals for multiple conditions.
}
\examples{
\donttest{
data(imputed_mtcars)
outcome <- 'qsec'
predictors <- c('cyl', 'disp', 'hp', 'drat', 'wt', 'vs', 'am', 'gear', 'carb','mpg')
imputation <- '.imp'
results <- ssvs_mi(data = imputed_mtcars, y = outcome, x = predictors, imp = imputation)
summary_est <- summary.est(results)
summary_mip <- summary.mip(results)
plot.est(summary_est)
plot.mip(summary_mip)
}
}
