% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MI.R
\name{plot.mip}
\alias{plot.mip}
\title{Creates a plot for marginal inclusion probabilities (MIP) optional thresholds for highlighting significant predictors.}
\usage{
\method{plot}{mip}(data, threshold = 0.5, legend = TRUE, title = NULL, color = TRUE)
}
\arguments{
\item{data}{A data frame containing MIP summary statistics. Must include columns:
\describe{
\item{Variables}{The predictor variables.}
\item{avg.mip}{The average MIP values across imputations or replications.}
\item{min}{The minimum MIP values.}
\item{max}{The maximum MIP values.}
}}

\item{threshold}{A numeric value (between 0 and 1) specifying the MIP threshold to highlight significant predictors.
Defaults to 0.5.}

\item{legend}{Logical indicating whether to include a legend for the threshold. Defaults to \code{TRUE}.}

\item{title}{A character string specifying the plot title. Defaults to \code{"Multiple Inclusion Probability for SSVS-MI"}.}

\item{color}{Logical indicating whether to use color to highlight thresholds. Defaults to \code{TRUE}.}
}
\value{
A \code{ggplot2} object representing the plot of MIP with thresholds.
}
\description{
Creates a plot for marginal inclusion probabilities (MIP) optional thresholds for highlighting significant predictors.
}
\examples{
\donttest{
# Example using imputed data
data(imputed_mtcars)
outcome <- "qsec"
predictors <- c("cyl", "disp", "hp", "drat", "wt", "vs", "am", "gear", "carb", "mpg")
imputation <- ".imp"
results <- ssvs_mi(data = imputed_mtcars, y = outcome, x = predictors, imp = imputation)
summary_MI <- summary.mi(results)
plot.mip(summary_MI)
}
}
