% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ssvs}
\alias{plot.ssvs}
\title{Plot results of an SSVS model}
\usage{
\method{plot}{ssvs}(x, y, MIP_threshold = 0.5, ...)
}
\arguments{
\item{x}{An SSVS result object obtained from \code{\link[=ssvs]{ssvs()}}}

\item{y}{The name of the dependent variable}

\item{MIP_threshold}{An MIP threshold to show on the plot}

\item{...}{Ignored}
}
\value{
Creates a plot of the inclusion probabilities by variable
}
\description{
Plot results of an SSVS model
}
\examples{
outcome <- "qsec"
predictors <- c("cyl", "disp", "hp", "drat", "wt", "vs", "am", "gear", "carb", "mpg")
results <- ssvs(x = predictors, y = outcome, data = mtcars, plot = FALSE)
plot(results, outcome)
}
