% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SSVS.R
\name{plot_SSVS}
\alias{plot_SSVS}
\title{SSVS plot function}
\usage{
plot_SSVS(y, ssvs.results, MIP_threshold = 0.5)
}
\arguments{
\item{y}{The name of the dependent variable}

\item{ssvs.results}{The result list from running the SSVS function}

\item{MIP_threshold}{An MIP threshold to show on the plot}
}
\value{
Creates a plot of the inclusion probabilities by variable
}
\description{
SSVS plot function
}
\examples{
outcome <- "qsec"
predictors <- c("cyl", "disp", "hp", "drat", "wt", "vs", "am", "gear", "carb", "mpg")
results <- SSVS(x = predictors, y = outcome, data = mtcars, plot = FALSE)
plot_SSVS(outcome,results)
}
