% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MI.R
\name{plot_ssvs_est}
\alias{plot_ssvs_est}
\title{Plot SSVS-MI Estimates with Confidence Intervals}
\usage{
plot_ssvs_est(
  data,
  ty = NULL,
  pal = NULL,
  Condition = NULL,
  cond = TRUE,
  title = NULL
)
}
\arguments{
\item{data}{A data frame containing the summary statistics for SSVS results. Must include columns
\code{Variables}, \code{avg.beta}, \code{min}, and \code{max} with option \code{Condition}.}

\item{ty}{A character vector specifying the order of conditions in the plot.}

\item{pal}{A character vector of colors corresponding to the conditions.}
}
\value{
A \code{ggplot2} object representing the SSVS estimates plot.
}
\description{
This function creates a plot of SSVS-MI estimates with confidence intervals for multiple conditions.
}
\examples{
\donttest{
data(example_data)
outcome <- "yMCAR40"
predictors <- c("xMCAR40_1", "xMCAR40_2", "xMCAR40_3", "xMCAR40_4", "xMCAR40_5")
results <- SSVS_MI(data = example_data, y = outcome, x = predictors)
summary_stats <- summary_MI(results)
plot_ssvs_est(summary_stats)
}
}
