% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_MI.R
\name{print.ssvs_mi_summary}
\alias{print.ssvs_mi_summary}
\title{Print the summary of ssvs_mi}
\usage{
\method{print}{ssvs_mi_summary}(x, ...)
}
\description{
Print the summary of ssvs_mi
}
\keyword{internal}
