% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.ssvs_summary}
\alias{print.ssvs_summary}
\title{Print the summary of an SSVS model}
\usage{
\method{print}{ssvs_summary}(x, ...)
}
\description{
Print the summary of an SSVS model
}
\keyword{internal}
