% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_MI.R
\name{summary.est}
\alias{summary.est}
\title{Calculate Summary Statistics for SSVS-MI Results}
\usage{
\method{summary}{est}(data)
}
\arguments{
\item{x}{An SSVS result object or a compatible data frame containing model estimates.}
}
\value{
A data frame with variables, average, minimum, and maximum for beta coefficients and MIP
}
\description{
Computes summary statistics (average, minimum, and maximum) for beta coefficients and MIP
from an SSVS result object.
}
\examples{
\donttest{
data(imputed_mtcars)
outcome <- 'qsec'
predictors <- c('cyl', 'disp', 'hp', 'drat', 'wt', 'vs', 'am', 'gear', 'carb','mpg')
imputation <- '.imp'
results <- ssvs_mi(data = imputed_mtcars, y = outcome, x = predictors, imp = imputation)
summary_est <- summary.est(results)
print(summary_est)
summary_mip <- summary.mip(results)
print(summary_mip)
}
}
