% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_MI.R
\name{summary.mi}
\alias{summary.mi}
\title{Calculate Summary Statistics for SSVS-MI Results}
\usage{
\method{summary}{mi}(data)
}
\arguments{
\item{x}{An SSVS result object or a compatible data frame containing model estimates.}
}
\value{
A data frame with variables, average, minimum, and maximum for beta, MIP and nonzero beta
}
\description{
Computes summary statistics (average, minimum, and maximum) for beta coefficients, MIP and
average nonzero beta coefficients from an SSVS result object.
}
\examples{
\donttest{
data(imputed_mtcars)
outcome <- 'qsec'
predictors <- c('cyl', 'disp', 'hp', 'drat', 'wt', 'vs', 'am', 'gear', 'carb','mpg')
imputation <- '.imp'
results <- ssvs_mi(data = imputed_mtcars, y = outcome, x = predictors, imp = imputation)
summary_MI <- summary.mi(results)
print(summary_MI)
}
}
