% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MI.R
\name{plot.est}
\alias{plot.est}
\title{Plot SSVS-MI Estimates and Marginal Inclusion Probabilities (MIP)}
\usage{
\method{plot}{est}(data, title = NULL)
}
\arguments{
\item{data}{A data frame containing summary statistics for SSVS results. Must include columns:
\describe{
\item{Variables}{The predictor variables.}
\item{avg.beta}{The average beta coefficients across imputations or replications.}
\item{min}{The minimum beta coefficients.}
\item{max}{The maximum beta coefficients.}
}}

\item{title}{A character string specifying the plot title. Defaults to \code{"SSVS-MI estimates"}.}
}
\value{
A \code{ggplot2} object representing the plot of SSVS estimates.
}
\description{
This function creates a plot of SSVS-MI estimates with minimum and maximum.
}
\examples{
\donttest{
data(imputed_mtcars)
outcome <- 'qsec'
predictors <- c('cyl', 'disp', 'hp', 'drat', 'wt', 'vs', 'am', 'gear', 'carb','mpg')
imputation <- '.imp'
results <- ssvs_mi(data = imputed_mtcars, y = outcome, x = predictors, imp = imputation)
summary_est <- summary.est(results)
plot.est(summary_est)
}
}
