% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MI.R
\name{plot.est}
\alias{plot.est}
\title{Plot results of SSVS across Multiple Imputations}
\usage{
\method{plot}{est}(data, title = NULL)
}
\arguments{
\item{data}{A data frame containing summary statistics for SSVS results. Must include columns:
\describe{
\item{Variable}{The predictor variables.}
\item{Avg Beta}{The average beta coefficients across imputations or replications.}
\item{Min Beta}{The minimum beta coefficients.}
\item{Max Beta}{The maximum beta coefficients.}
}}

\item{title}{A character string specifying the plot title. Defaults to \code{"SSVS-MI estimates"}.}
}
\value{
A \code{ggplot2} object representing the plot of SSVS estimates.
}
\description{
This function creates a plot of SSVS coefficient estimates with the average, minimum, and maximum across imputations.
}
\examples{
\donttest{
data(imputed_mtcars)
outcome <- 'qsec'
predictors <- c('cyl', 'disp', 'hp', 'drat', 'wt', 'vs', 'am', 'gear', 'carb','mpg')
imputation <- '.imp'
results <- ssvs_mi(data = imputed_mtcars, y = outcome, x = predictors, imp = imputation)
summary_MI <- summary.mi(results)
plot.est(summary_MI)
}
}
