% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSVS.R
\name{SSVS}
\alias{SSVS}
\title{SSVS Function}
\usage{
SSVS(
  y,
  x,
  runs = 20000,
  burn = 5000,
  update = 1000,
  a1 = 0.01,
  b1 = 0.01,
  prec.beta = 0.1,
  inprob = 0.5
)
}
\arguments{
\item{y}{The response variable}

\item{x}{The set of predictor variables}

\item{runs}{test}

\item{burn}{test}

\item{update}{test}

\item{a1}{test}

\item{b1}{test}

\item{prec.beta}{test}

\item{inprob}{test}
}
\value{
Returns a list
}
\description{
SSVS Function
}
