% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSVS.R
\name{SSVS}
\alias{SSVS}
\title{SSVS Function}
\usage{
SSVS(
  x,
  y,
  data,
  runs = 20000,
  burn = 5000,
  a1 = 0.01,
  b1 = 0.01,
  prec.beta = 0.1,
  inprob = 0.5
)
}
\arguments{
\item{x}{The set of predictor variables}

\item{y}{The continuous response variable}

\item{data}{The dataframe used to extract predictors and response values}

\item{runs}{Total number of iterations (including burn-in). Results are based on
the Total - Burn-in iterations.}

\item{burn}{Number of burn-in iterations. Burn-in iterations are discarded
warmup iterations used to achieve MCMC convergence. You may increase the number
of burn-in iterations if you are having convergence issues.}

\item{a1}{Prior parameter for Gamma(a,b) distribution on the precision (1/variance)
residual variance.}

\item{b1}{Prior parameter for Gamma(a,b) distribution on the precision (1/variance)
residual variance.}

\item{prec.beta}{Prior precision (1/variance) for beta coefficients}

\item{inprob}{Prior inclusion probability value, which applies to all predictors.
The prior inclusion probability reflects the prior belief that each predictor
should be included in the model. A prior inclusion probability of .5 reflects
the belief that each predictor has an equal probability of being included or
excluded. Note that a value of .5 also implies a prior belief that the true model
contains half of the candidate predictors. The prior inclusion probability will
influence the magnitude of the marginal inclusion probabilities (MIPs), but the
relative pattern of MIPs is expected to remain fairly consistent, see Bainter et al.
(2020) for more information.}
}
\value{
Returns a list
}
\description{
This function performs SSVS for continuous outcomes using a basic gibbs sampler
}
\examples{
ex1 <- SSVS(Mean_Unpleasantness~Age+BFI_Agreeableness+BFI_Conscientiousness+BFI_Extraversion
+BFI_Neuroticism+BFI_Openness+PANAS_Negative_Prescan+PANAS_Positive_Prescan
+PANAS_Trait_Negative+PANAS_Trait_Positive,data=dat)


}
