% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_SSVS.R
\name{summary_SSVS}
\alias{summary_SSVS}
\title{SSVS summary function}
\usage{
summary_SSVS(ssvs.results)
}
\arguments{
\item{ssvs.results}{The result list from running the SSVS function}
}
\value{
Returns a dataframe with results
}
\description{
Function to summarize results from SSVS including marginal inclusion probabilities,
Bayesian model averaged parameter estimates, and 
95% highest posterior density credible intervals
}
